+++ lib/storage/src/content_manager/consensus/mod.rs
@@ -1,3 +1,4 @@
+pub mod applied_log;
 pub mod consensus_wal;
 pub mod entry_queue;
 pub mod operation_sender;
